api/requirements.txt
Django
django-filters
django-private-chat
djangorestframework
django-rest-auth
django-allauth
django-session-timeout
